carbonx/oracle/requirements.txt
fastapi
uvicorn
python-dotenv
web3
pydantic
